refactor(Association) : refactor code, create adapter, fragment and util package
