+++ app/src/main/java/fr/paris10/projet/assogenda/assogenda/ui/fragment/CreateAssociationFragment.java
@@ -1,4 +1,4 @@
-package fr.paris10.projet.assogenda.assogenda.ui.activites;
+package fr.paris10.projet.assogenda.assogenda.ui.fragment;
 
 import android.app.Activity;
 import android.content.Context;
