+++ app/src/main/res/layout/fragment_create_association.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fr.paris10.projet.assogenda.assogenda.ui.activites.CreateAssociationFragment">
+    tools:context="fr.paris10.projet.assogenda.assogenda.ui.fragment.CreateAssociationFragment">
 
     <LinearLayout
         android:orientation="vertical"
